pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Diego Garcia", email = "diegovianagarcia@gmail.com"},
]
dependencies = [
    "djangorestframework>=3.13.1",
]
requires-python = ">=3.10.2"
license = {text = "MIT"}
[project.optional-dependencies]

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=1.6.0",
    "pylint>=2.14.5",
    "pytest>=7.1.2",
]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"

[tool.pyright]
include = ["src"]
exclude = [
    "**/__pycache__",
    "__pypackages__",
    ".pytest_cache"
]



